Ok let's see if I fixed the img inlining in markdown exporting.
The noise example post processor takes the full pathname of the
written image, then because it is actually getting a string version
of it including the #P" and " on the other end as the LITERAL string
from org mode, we pick it apart, find the cwd of the emacs process
and remove the prefix to get the relative pathname.

This is a total hack, but good enough for now.